website_structure_2/requirements.txt
Flask
requests
Pillow
osmnx
